+++ suite-common/test-utils/src/extraDependenciesCommonMock.ts
@@ -86,7 +86,7 @@ export const extraDependenciesCommonMock: ExtraDependencies = {
             ),
         connectInitSettings,
         connectInitHooks: { deviceEvent: {}, uiEvent: {} },
-        mapDebugTransports: (debugTransports: readonly unknown[] | undefined) =>
+        mapDebugTransports: ({ debugTransports }) =>
             debugTransports as ReturnType<ExtraDependencies['services']['mapDebugTransports']>,
         migrateSuiteSyncLabelsForRbfTransaction: () => Promise.resolve([[], []]),
     },